refactor(Viewer): Move function to helper file
This function being exported, in order to avoid any BC,
we continue to export it from the component file
